+++ src/main/kotlin/tech/relaycorp/relaynet/wrappers/x509/Certificate.kt
@@ -1,4 +1,4 @@
-package tech.relaycorp.relaynet
+package tech.relaycorp.relaynet.wrappers.x509
 
 import java.security.KeyPair
 import java.security.KeyPairGenerator
